+++ packages/zowe-explorer/__tests__/__mocks__/@zowe/imperative.ts
@@ -269,7 +269,9 @@ export class ProfileInfo {
     ): any {
         return;
     }
+}
 
+export class ConfigUtils {
     public static getZoweDir(): string {
         const defaultHome = path.join(os.homedir(), ".zowe");
         if (ImperativeConfig.instance.loadedConfig?.defaultHome !== defaultHome) {
@@ -282,7 +284,6 @@ export class ProfileInfo {
         return ImperativeConfig.instance.cliHome;
     }
 }
-
 export class ImperativeError extends Error {
     private msg: string;
     constructor(public mDetails: IImperativeError) {
@@ -425,6 +426,22 @@ export class ConvertV1Profiles {
             profilesConverted: [],
             profilesFailed: [],
             v1ScsPluginName: undefined,
+        };
+    }
+}
+
+export class ZoweUserEvents {
+    public static readonly ON_VAULT_CHANGED = "onVaultChanged";
+}
+export class ZoweSharedEvents {
+    public static readonly ON_CREDENTIAL_MANAGER_CHANGED = "onCredentialManagerChanged";
+}
+
+export class EventOperator {
+    public static getWatcher() {
+        return {
+            subscribeUser: () => ({ close: () => {} }),
+            subscribeShared: () => ({ close: () => {} }),
         };
     }
 }