+++ packages/zowe-explorer/__tests__/__unit__/trees/dataset/DatasetActions.unit.test.ts
@@ -240,7 +240,7 @@ describe("Dataset Actions Unit Tests - Function createMember", () => {
             session: blockMocks.session,
         });
         const nonFavoriteLabel = parent.label;
-        parent.label = `${parent.label}`;
+        parent.label = `${parent.label?.toString()}`;
         parent.contextValue = Constants.DS_PDS_CONTEXT + Constants.FAV_SUFFIX;
 
         const mySpy = mocked(vscode.window.showInputBox).mockResolvedValue("testMember");
@@ -308,7 +308,7 @@ describe("Dataset Actions Unit Tests - Function refreshPS", () => {
         expect(mocked(vscode.commands.executeCommand)).not.toHaveBeenCalled();
     });
     it("Checking failed attempt to refresh PS dataset (not found exception)", async () => {
-        const globalMocks = createGlobalMocks();
+        createGlobalMocks();
         const blockMocks = createBlockMocksShared();
         const node = new ZoweDatasetNode({
             label: "HLQ.TEST.AFILE7",
@@ -335,7 +335,7 @@ describe("Dataset Actions Unit Tests - Function refreshPS", () => {
         blockMocks.fetchDsAtUri.mockRejectedValueOnce(Error("not found"));
 
         await DatasetActions.refreshPS(child);
-        expect(mocked(Gui.showMessage)).toHaveBeenCalledWith(`Unable to find file ${parent.label}(${child.label})`);
+        expect(mocked(Gui.showMessage)).toHaveBeenCalledWith(`Unable to find file ${parent.label?.toString()}(${child.label?.toString()})`);
     });
     it("Checking favorite empty PDS refresh", async () => {
         createGlobalMocks();
@@ -441,7 +441,7 @@ describe("Dataset Actions Unit Tests - Function deleteDatasetPrompt", () => {
 
         mocked(vscode.window.withProgress).mockImplementation((progLocation, callback) => {
             const progress = {
-                report: (message) => {
+                report: (_message) => {
                     return;
                 },
             };
@@ -476,7 +476,9 @@ describe("Dataset Actions Unit Tests - Function deleteDatasetPrompt", () => {
 
         await DatasetActions.deleteDatasetPrompt(blockMocks.testDatasetTree);
 
-        expect(mocked(Gui.showMessage)).toHaveBeenCalledWith(`The following 1 item(s) were deleted: ${blockMocks.testDatasetNode.getLabel()}`);
+        expect(mocked(Gui.showMessage)).toHaveBeenCalledWith(
+            `The following 1 item(s) were deleted: ${blockMocks.testDatasetNode.getLabel().toString()}`
+        );
     });
 
     it("Should delete one member", async () => {
@@ -492,7 +494,8 @@ describe("Dataset Actions Unit Tests - Function deleteDatasetPrompt", () => {
         await DatasetActions.deleteDatasetPrompt(blockMocks.testDatasetTree);
 
         expect(mocked(Gui.showMessage)).toHaveBeenCalledWith(
-            `The following 1 item(s) were deleted: ${blockMocks.testMemberNode.getParent().getLabel()}(${blockMocks.testMemberNode.getLabel()})`
+            `The following 1 item(s) were deleted: ` +
+                `${blockMocks.testMemberNode.getParent().getLabel().toString()}(${blockMocks.testMemberNode.getLabel().toString()})`
         );
     });
 
@@ -507,7 +510,9 @@ describe("Dataset Actions Unit Tests - Function deleteDatasetPrompt", () => {
 
         await DatasetActions.deleteDatasetPrompt(blockMocks.testDatasetTree);
 
-        expect(mocked(Gui.showMessage)).toHaveBeenCalledWith(`The following 1 item(s) were deleted: ${blockMocks.testVsamNode.getLabel()}`);
+        expect(mocked(Gui.showMessage)).toHaveBeenCalledWith(
+            `The following 1 item(s) were deleted: ${blockMocks.testVsamNode.getLabel().toString()}`
+        );
     });
 
     it("Should delete one migrated dataset", async () => {
@@ -521,7 +526,9 @@ describe("Dataset Actions Unit Tests - Function deleteDatasetPrompt", () => {
 
         await DatasetActions.deleteDatasetPrompt(blockMocks.testDatasetTree);
 
-        expect(mocked(Gui.showMessage)).toHaveBeenCalledWith(`The following 1 item(s) were deleted: ${blockMocks.testMigrNode.getLabel()}`);
+        expect(mocked(Gui.showMessage)).toHaveBeenCalledWith(
+            `The following 1 item(s) were deleted: ${blockMocks.testMigrNode.getLabel().toString()}`
+        );
     });
 
     it("Should delete two datasets", async () => {
@@ -536,7 +543,8 @@ describe("Dataset Actions Unit Tests - Function deleteDatasetPrompt", () => {
         await DatasetActions.deleteDatasetPrompt(blockMocks.testDatasetTree);
 
         expect(mocked(Gui.showMessage)).toHaveBeenCalledWith(
-            `The following 2 item(s) were deleted: ${blockMocks.testDatasetNode.getLabel()}, ${blockMocks.testVsamNode.getLabel()}`
+            `The following 2 item(s) were deleted: ` +
+                `${blockMocks.testDatasetNode.getLabel().toString()}, ${blockMocks.testVsamNode.getLabel().toString()}`
         );
     });
 
@@ -551,7 +559,9 @@ describe("Dataset Actions Unit Tests - Function deleteDatasetPrompt", () => {
 
         await DatasetActions.deleteDatasetPrompt(blockMocks.testDatasetTree);
 
-        expect(mocked(Gui.showMessage)).toHaveBeenCalledWith(`The following 1 item(s) were deleted: ${blockMocks.testDatasetNode.getLabel()}`);
+        expect(mocked(Gui.showMessage)).toHaveBeenCalledWith(
+            `The following 1 item(s) were deleted: ${blockMocks.testDatasetNode.getLabel().toString()}`
+        );
     });
 
     it("Should delete a favorited data set", async () => {
@@ -567,7 +577,7 @@ describe("Dataset Actions Unit Tests - Function deleteDatasetPrompt", () => {
 
         expect(mocked(Gui.warningMessage)).toHaveBeenCalledWith(
             `Are you sure you want to delete the following 1 item(s)?\nThis will permanently remove these data sets and/or members from your ` +
-                `system.\n\n ${blockMocks.testFavoritedNode.getLabel()}`,
+                `system.\n\n ${blockMocks.testFavoritedNode.getLabel().toString()}`,
             { items: ["Delete"], vsCodeOpts: { modal: true } }
         );
     });
@@ -585,7 +595,7 @@ describe("Dataset Actions Unit Tests - Function deleteDatasetPrompt", () => {
 
         expect(mocked(Gui.warningMessage)).toHaveBeenCalledWith(
             `Are you sure you want to delete the following 1 item(s)?\nThis will permanently remove these data sets and/or members from your ` +
-                `system.\n\n ${blockMocks.testFavoritedNode.getLabel()}(${blockMocks.testFavMemberNode.getLabel()})`,
+                `system.\n\n ${blockMocks.testFavoritedNode.getLabel().toString()}(${blockMocks.testFavMemberNode.getLabel().toString()})`,
             { items: ["Delete"], vsCodeOpts: { modal: true } }
         );
     });
@@ -616,7 +626,8 @@ describe("Dataset Actions Unit Tests - Function deleteDatasetPrompt", () => {
         await DatasetActions.deleteDatasetPrompt(blockMocks.testDatasetTree);
 
         expect(mocked(Gui.showMessage)).toHaveBeenCalledWith(
-            `The following 2 item(s) were deleted: ${blockMocks.testDatasetNode.getLabel()}, ${blockMocks.testFavoritedNode.getLabel()}`
+            `The following 2 item(s) were deleted: ` +
+                `${blockMocks.testDatasetNode.getLabel().toString()}, ${blockMocks.testFavoritedNode.getLabel().toString()}`
         );
     });
 
@@ -642,7 +653,8 @@ describe("Dataset Actions Unit Tests - Function deleteDatasetPrompt", () => {
 
         await DatasetActions.deleteDatasetPrompt(blockMocks.testDatasetTree, blockMocks.testMemberNode);
         expect(mocked(Gui.showMessage)).toHaveBeenCalledWith(
-            `The following 1 item(s) were deleted: ${blockMocks.testMemberNode.getParent().getLabel()}(${blockMocks.testMemberNode.getLabel()})`
+            `The following 1 item(s) were deleted: ` +
+                `${blockMocks.testMemberNode.getParent().getLabel().toString()}(${blockMocks.testMemberNode.getLabel().toString()})`
         );
     });
 });
@@ -746,7 +758,7 @@ describe("Dataset Actions Unit Tests - Function deleteDataset", () => {
         mocked(vscode.window.showQuickPick).mockResolvedValueOnce("Delete" as any);
         jest.spyOn(DatasetFSProvider.instance, "delete").mockRejectedValueOnce(Error("not found"));
         await expect(DatasetActions.deleteDataset(node, blockMocks.testDatasetTree)).rejects.toThrow("not found");
-        expect(mocked(Gui.showMessage)).toHaveBeenCalledWith("Unable to find file " + node.label);
+        expect(mocked(Gui.showMessage)).toHaveBeenCalledWith("Unable to find file " + node.label?.toString());
     });
     it("Checking common PS dataset failed deletion attempt", async () => {
         createGlobalMocks();
@@ -2967,7 +2979,7 @@ describe("Dataset Actions Unit Tests - Function confirmJobSubmission", () => {
 });
 
 describe("Dataset Actions Unit Tests - Function getDsTypePropertiesFromWorkspaceConfig", () => {
-    it("Should use use local JCL doc name for confirmJobSubmission", async () => {
+    it("Should use use local JCL doc name for confirmJobSubmission", () => {
         createGlobalMocks();
         const options = createWorkspaceConfiguration();
         // const opt = DatasetActions.getDsTypePropertiesFromWorkspaceConfig(options);